feat: split-off `AutoinstallDirectPage`
