+++ packages/ubuntu_provision/assets/whitelabel.yml
@@ -15,6 +15,8 @@ pages:
     image: "assets/images/mascot.svg"
   autoinstall:
     image: "assets/images/disk.svg"
+  autoinstall-direct:
+    image: "assets/images/disk.svg"
   source-selection:
     image: "assets/images/applications.svg"
   codecs-and-drivers:
